tests

+++ Transactions.java
@@ -1059,7 +1059,7 @@ public class Transactions {
                                             "WHERE h_t_id = %s", hold_qty+needed_qty, entry.get("h_t_id"));
                             st.executeUpdate(trFrame2_5a);
                             hStats.incWriteOp();
-                            sell_value 	+= needed_qty*Integer.parseInt(entry.get("h_price").toString());
+                            sell_value 	+= needed_qty*Double.parseDouble(entry.get("h_price").toString());
                             buy_value 	+= needed_qty*trade_price;
                             needed_qty 	= 0;
                             continue;
